Move GPT algos to individual files
